backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
shapely==2.0.6
networkx==3.4.2
opencv-python-headless==4.10.0.84
anthropic==0.40.0
python-multipart==0.0.12
supabase==2.9.0
pdfplumber==0.11.4
ezdxf==1.3.4
python-dotenv==1.0.1
pymupdf==1.24.14
trimesh==4.11.5
scipy==1.17.1
